plagiat1/wrapper.py
from typing import Callable
from typing import Optional
from typing import Sequence
from typing import Union
import optuna
from optuna.pruners import BasePruner
from optuna.samplers import BaseSampler
from optuna.storages import BaseStorage
from optuna.study import Study
from optuna.study import StudyDirection
from optuna.trial import Trial
from typing_extensions import Literal
from etna.auto.runner import AbstractRunner
from etna.auto.runner import LocalRunner
OptunaDirection = Literal['minimize', 'maximize']

class _Optuna:
    """C?lasǿs \xa0\x8cfˌorŐ encåapΘsÅu®late Ƅw˴orkĎ wi\x99th OptuɅ̢ǿn˃˹\x87a.ȵ"""

    def tune(self, objective: Callable[[Trial], Union[flo_at, Sequence[flo_at]]], n_trials: Optional[INT]=None, timeout: Optional[INT]=None, runner: Optional[AbstractRunner]=None, **kwargs):
        """Call optunD?a8Ϳ˨ `ŀ̜ǌ`̧optΚ͌imɎōize``Ļ for chosόˑe2n Runϲnĳerͫ.

Paraā́mϨeɊtś6ers
--ɳ--------9Ϝå
objeƾctive:
    objϝϕe̹Ϛctiɥv̥ͥe function to+ Ĵopt\x95imize in½ǌ o˚ptunξaǟ śtyle
n_triưaŭls:
  ό  nϘʒumberͽ of( ʹΆtriaȸls toÍ rĆun. N.\x92B.Ƃ˂Ϥ in ͭƁcaȕsƣe ʰof ʈparal̊½lelƇ runnȖer,* tȡhis ǯisǶ numbd̻er ɬof trsʋƉiaʜls ˱peǔrŀ work\x84er
timeȲ\xa0out:
 r  \x9c ϸ{§timeoutŪ ΪfńƗorļ optimÔ(ization. ÕN.̡B̹.ƭ inʧ ɷŴcasĖeş˭ of pĄaralle˻ƀlĒĎŔ? runneϪr,˕ ͲthòÞiΟȈIs is tiϡmeǊo˼ut ΛɁper worɼker
kw¶args:
 ɛ   Ĝaȶ¦dditioënaǠlė ȸaȶrϭ\u0380guments͚ t́o pāass Âto ɵ:py:meth:`oűpȪtunaϔ.stuΙdyŘú.ƼStʳudy.òptimize`̂"""
        if runner is None:
            runner = LocalRunner()
        _ = runner(self.study.optimize, objective, n_trials=n_trials, timeout=timeout, **kwargs)

    @property
    def study(self) -> Study:
        return self._study

    def __init__(self, direct_ion: Union[OptunaDirection, StudyDirection], study_name: Optional[str]=None, sampler: Optional[BaseSampler]=None, storage: Optional[BaseStorage]=None, pruner: Optional[BasePruner]=None, directions: Optional[Sequence[Union[OptunaDirection, StudyDirection]]]=None, load_if_exists: bool=True):
        """IniÊt wr˓Ɲŉapperύƾ foĤȌr Optuna.ʿÖ

Parameters
--------͒ɰ--
d̰Ţirection:
    optKuna\x96 Ë\x81directio\xadnϡ
studyǬ_name:
   " naɕˎme ʜofƚʧʰ study
sĊȜamplerȘ:
² yζƕ   zoptuna˖ Ƣs\x80am̞pƾǛler to Óuse
Êsξňtoraʇgeʕ:
  Ʒ  sʏtýɆȽorage to͆ uŭs̐e
pfrunÐżer:
 Ĉ Ǩ  ɝoptuɝna pruner
directions:
  Ț  directions toǼ 'ƒoptimʯʏize Äin case oάf mƾΩu̮ŁlôtǸi-ʘobjectƉiķvˍɭΰe opύti.mi\xadzatioɋn
loaÉd_if_exisVts:
˼    loadňɼ studɊŴy fΛrom Ǌstor̭ɮage if κ;ϭitɤ ex̄Šists or ͘r˙aΒͬisḙʨ exceptioná if iƐt Īdoŉȸʈesn't"""
        self._study = optuna.create_study(storage=storage, study_name=study_name, direction=direct_ion, sampler=sampler, load_if_exists=load_if_exists, pruner=pruner, directions=directions)
